Delete README.md
Why it doesn't make sense? The question is correct
I think purpose of the question was to examine Common Programming Error 16.5 (in eight edition):

Use caution when throwing the result of a conditional expression ( ?: )—promotion rules
could cause the value to be of a type different from the one expected. For example, when
throwing an int or a double from the same conditional expression, the int is promoted
to a double . So, a catch handler that catches an int would never execute based on such
a conditional expression.